+++ src_java/src/main/java/ch/konnexions/orabench/utils/Setup.java
@@ -142,6 +142,7 @@ public class Setup {
   /**
    * Creates a new result file if none exists yet.
    */
+  @SuppressWarnings("resource")
   private void createResultFile() {
     if (isDebug) {
       logger.debug("Start");